move img utils to activity util dir
